list files separation
