+++ maslevtsov.stanislav/S1/fwd_list/fwd_list_node.hpp
