+++ maslevtsov.stanislav/S1/fwd_list_operations.hpp
@@ -3,7 +3,7 @@
 
 #include <cstddef>
 #include <string>
-#include "fwd_list.hpp"
+#include "fwd_list/definition.hpp"
 
 namespace maslevtsov {
   using list_t = maslevtsov::FwdList< unsigned long long >;
